requirements.txt
watchdog

portalocker

pytest
Pillow
tensorflow==2.11
nsfw_detector
deepdanbooru
